requirements.txt
firebase_admin
flask
flask-bcrypt
flask-cors
PyJWT
python-dotenv
requests